+++ tests/files/no-parent-barrel-import/foo/bar.js
@@ -1,0 +1,1 @@
+// Used in `no-parent-barrel-import` tests