+++ tests/files/no-parent-barrel-import/index.js
@@ -1,0 +1,2 @@
+// Used in `no-parent-barrel-import` tests
+